+++ tests/shapes/test_shell.py
@@ -1,7 +1,8 @@
 import unittest
 import volmdlr
-from volmdlr import shapes, wires, surfaces, faces
+from volmdlr import curves, shapes, wires, surfaces, faces
 from OCP.BRepPrimAPI import BRepPrimAPI_MakeBox
+from OCP.TopoDS import TopoDS_Shell
 
 
 class TestShell(unittest.TestCase):
@@ -65,6 +66,23 @@ class TestShell(unittest.TestCase):
         self.assertAlmostEqual(shell.primitives[0].area(), length * height)
         self.assertAlmostEqual(shell.primitives[1].area(), width * height)
         self.assertFalse(shell.is_closed)
+
+    def test_loft(self):
+        diameter = 0.3
+        circle1 = curves.Circle3D(frame=volmdlr.OXYZ, radius=diameter / 2)
+        circle2 = curves.Circle3D(
+            frame=volmdlr.Frame3D(volmdlr.Point3D(0.3, 0.0, 0.5), volmdlr.Y3D, volmdlr.Z3D, volmdlr.X3D),
+            radius=circle1.radius / 2)
+        circle3 = curves.Circle3D(
+            frame=volmdlr.Frame3D(volmdlr.Point3D(0.6, 0.0, 0.3), volmdlr.Y3D, volmdlr.X3D, -volmdlr.Z3D),
+            radius=circle1.radius * 0.6)
+        sections = [wires.Contour3D.from_circle(circle1), wires.Contour3D.from_circle(circle2),
+                    wires.Contour3D.from_circle(circle3)]
+        loft = shapes.Shell.make_loft(sections=sections, ruled=True, name="loft")
+        self.assertEqual(loft.name, "loft")
+        self.assertEqual(len(loft.primitives), 4)
+        self.assertFalse(loft.is_closed)
+        self.assertIsInstance(loft.wrapped, TopoDS_Shell)
 
 
 if __name__ == '__main__':