+++ volmdlr/from_ocp.py
@@ -223,10 +223,23 @@ def bsplinecurve3d_from_ocp(cls, curve):
     return cls(curve.Degree(), control_points, multiplicities, knots, weigths)
 
 
+def beziercurve3d_from_ocp(cls, curve):
+    """
+    Instanciates a volmdlr BezierCurve3D, from occt object.
+
+    :param cls: volmdlr class to be instanciated.
+    :param curve: OCCT curve.
+    :return: volmdlr BezierCurve3D.
+    """
+    control_points = [volmdlr.Point3D(point.X(), point.Y(), point.Z()) for point in curve.Poles()]
+    return cls(curve.Degree(), control_points)
+
+
 OCCT_TO_VOLMDLR = {"Geom_Line": line3d_from_ocp,
                    "Geom_Circle": circle3d_from_ocp,
                    "Geom_Ellipse": ellipse3d_from_ocp,
                    "Geom_BSplineCurve": bsplinecurve3d_from_ocp,
+                   "Geom_BezierCurve": beziercurve3d_from_ocp,
                    "Geom_Parabola": parabola3d_from_ocp,
                    "Geom_Hyperbola": hyperbola3d_from_ocp,
                    "Geom2d_Line": line2d_from_ocp,